utils and hooks exported. tooltip defaults added
