+++ src/components/table/value-cells/table-device-cell/TableDeviceCell.tsx
@@ -19,9 +19,8 @@ import printer from '../../../assets/devices/printer.svg'
 import hardDrive2 from '../../../assets/devices/hard-drive-2.svg'
 import TableBoxEllipses from '../table-box-ellipses'
 import {SVG} from '../../../svg'
-import {clipboard} from '../../../../utils/text'
 import {TooltipV2} from '../../../tooltip-v2'
-import {TOOLTIP_DEFAULTS} from '../../../../utils/constants'
+import {clipboard, TOOLTIP_DEFAULTS} from '../../../../utils'
 
 interface TableDeviceCellProps {
   device: any
@@ -51,7 +50,6 @@ export function TableDeviceCell({device, onClick, customStyle}: TableDeviceCellP
           )}
           {device.isMdmConnected && (
             <TooltipV2
-              id={'id'}
               trigger={
                 <SVG
                   path={remote}
