+++ src/hooks/use-color-from-word.ts
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import {getColorsFromWord} from '..'
+import {getColorsFromWord} from '../utils'
 
 export function useColorsFromWord(word: string): {darkerColor: string; lighterColor: string} {
   return React.useMemo(() => getColorsFromWord(word), [word])
