+++ src/utils/constants.ts
@@ -2,7 +2,6 @@ export const ZENADMIN_URL = 'https://www.zenadmin.ai'
 export const ZENADMIN_CONTACT = 'mailto:customersuccess@zenadmin.ai'
 export const TOOLTIP_DEFAULTS = {
   opacity: 1,
-  maxWidth: '260px',
   contentMaxLength: 250,
   portalId: 'root',
 }
